requirements.txt
streamlit>=1.34
pandas>=2.0
openpyxl>=3.1
xlrd>=2.0.1
Pillow>=10.0
requests>=2.31
pdfplumber>=0.11
